requirements.txt
# RAG dependencies
chromadb>=0.4.22
langchain>=0.1.11
langchain-community>=0.0.27
langchain-ollama>=0.1.1
langchain-core>=0.1.30
python-dotenv>=1.0.1
pypdf>=4.0.2
pydantic>=2.6.3
python-docx>=0.8.11
docx2txt>=0.8

# GraphRAG dependencies
networkx>=3.1
matplotlib>=3.7.2
pyvis>=0.3.2

# OCR dependencies
PyMuPDF>=1.19.0
pytesseract>=0.3.8
easyocr>=1.5.0
Pillow>=9.0.0
click>=8.0.0
numpy>=1.20.0
torch>=1.10.0
tqdm>=4.62.0

# Excel file support
pandas>=2.0.0
openpyxl>=3.1.0

# JSON and HTML support
beautifulsoup4>=4.10.0
lxml>=4.9.0
